pyproject.toml
[project]
name = "mycelium-onboarding"
version = "0.1.0"
description = "Mycelium onboarding system with XDG-compliant environment isolation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Mycelium Team", email = "team@mycelium.dev" }
]
keywords = ["onboarding", "xdg", "environment", "isolation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.12.1",
    "pyyaml>=6.0.3",
    "click>=8.1.0",
    "packaging>=23.0",
    "platformdirs>=4.0.0",  # XDG-compliant path resolution (Sprint 2)
    "inquirerpy>=0.3.4",
    "rich>=14.2.0",
    "jinja2>=3.1.0",
    "asyncpg>=0.29.0",  # PostgreSQL async driver for agent registry
    "fastapi>=0.104.0",  # FastAPI framework for Discovery API (Task 1.2)
    "uvicorn>=0.24.0",   # ASGI server for FastAPI
    "httpx>=0.25.0",     # HTTP client for MCP tools (Task 1.4)
    "fastmcp>=2.0.0",    # FastMCP SDK for MCP tools (Task 1.4, 1.8) - includes mcp
    "sentence-transformers>=2.2.0",  # NLP matching (Task 1.3)
    "jsonschema>=4.20.0",  # JSON schema validation (Task 1.6)
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-benchmark>=4.0.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.82.0",
    "httpx>=0.25.0",  # For testing FastAPI endpoints
]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]
# Optional dependencies for NLP features (Task 1.3)
nlp = [
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
]
# Optional dependencies for Discovery API (Task 1.2)
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",  # For form data support
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mycelium_onboarding"]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 120
indent-width = 4
extend-exclude = [
    "tests/docs/test_generated_snippets.py",  # Auto-generated, skip linting
    "tests/docs/snippets/",  # Extracted code snippets
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "RET",    # flake8-return
    "N",      # pep8-naming
    "D",      # pydocstyle"
]
ignore = [
    "D203",   # one-blank-line-before-class
    "D213",   # multi-line-summary-second-line
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",    # Allow unused args in tests
    "S101",   # Allow asserts in tests
    "D",      # Don't require docstrings in tests
]
"plugins/**/*.py" = [
    "D",      # Don't require docstrings in plugins (for now)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# Mypy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "uvicorn.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "packaging.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sentence_transformers.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "InquirerPy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "click.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jinja2.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rich.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "platformdirs.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
testpaths = ["tests"]
# REMOVED pythonpath line - causes sys.path issues
markers = [
    "benchmark: marks tests as benchmarks (deselect with '-m \"not benchmark\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "asyncio: marks tests as async tests",
]
asyncio_mode = "auto"

# Coverage configuration
[tool.coverage.run]
branch = true
source = ["mycelium_onboarding", "plugins"]
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "\\.\\.\\.",
]

[tool.coverage.html]
directory = "htmlcov"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.2.0",
    "types-pyyaml>=6.0.12.20250915",
    "types-jsonschema>=4.20.0",
    "pytest-mock>=3.15.1",
    "pre-commit>=4.3.0",
    "pandas>=2.3.3",
    "alembic>=1.17.1",
    "sqlalchemy>=2.0.44",
    "psycopg2-binary>=2.9.11",
]
